format the Calc test

+++ contracts/contracts/Calc.sol
@@ -3,15 +3,13 @@
 pragma solidity ^0.7.0;
 
 contract Calc {
-	address public owner;
-	function add(uint x, uint y) public returns(uint256)
-	{
-		return x+y;
-	}
- 
-	function sub(uint x, uint y) public returns(uint256)
-	{
-		return x - y;
-	}
- 
+    address public owner;
+
+    function add(uint256 x, uint256 y) public returns (uint256) {
+        return x + y;
+    }
+
+    function sub(uint256 x, uint256 y) public returns (uint256) {
+        return x - y;
+    }
 }
